+++ codewof/config/__init__.py
@@ -1,6 +1,6 @@
 """Configuration for Django system."""
 
-__version__ = "1.2.1"
+__version__ = "1.3.0"
 __version_info__ = tuple(
     [
         int(num) if num.isdigit() else num