+++ codewof/general/forms.py
@@ -7,8 +7,7 @@ from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Submit
 from captcha.fields import ReCaptchaField
 
-SUBJECT_TEMPLATE = "[CodeWOF] - {}"
-MESSAGE_TEMPLATE = "{}\n\n-----\nMessage sent from {}\n-----\ncodeWOF"
+MESSAGE_TEMPLATE = "{}\n\n-----\nMessage sent from {} <{}>\n"
 
 
 class ContactForm(forms.Form):
@@ -28,12 +27,12 @@ class ContactForm(forms.Form):
         from_email = self.cleaned_data['from_email']
         message = self.cleaned_data['message']
         mail_managers(
-            SUBJECT_TEMPLATE.format(subject),
+            subject,
             MESSAGE_TEMPLATE.format(message, name, from_email),
         )
         if self.cleaned_data.get('cc_sender'):
             send_mail(
-                SUBJECT_TEMPLATE.format(subject),
+                subject,
                 MESSAGE_TEMPLATE.format(message, name, from_email),
                 settings.DEFAULT_FROM_EMAIL,
                 [from_email],