+++ codewof/programming/admin.py
@@ -2,9 +2,18 @@
 
 from django.contrib import admin
 from django.contrib.auth import get_user_model
-from programming.models import Attempt, QuestionTypeDebugging
+from programming.models import Attempt, TestCaseAttempt
 
 User = get_user_model()
+
+
+class TestCaseAttemptInline(admin.TabularInline):
+    """Configuration for displaying test case attempts inline in admin."""
+
+    model = TestCaseAttempt
+    fields = ()
+    extra = 0
+    can_delete = False
 
 
 class AttemptAdmin(admin.ModelAdmin):
@@ -12,7 +21,14 @@ class AttemptAdmin(admin.ModelAdmin):
 
     list_display = ('datetime', 'question', 'profile', 'passed_tests')
     ordering = ('-datetime', )
+    inlines = (TestCaseAttemptInline, )
+
+    class Media:
+        """Add extra CSS rules."""
+
+        css = {
+            "all": ("css/admin.css",)
+        }
 
 
 admin.site.register(Attempt, AttemptAdmin)
-admin.site.register(QuestionTypeDebugging)