+++ codewof/programming/content/en/book-titles/solution.py
@@ -1,0 +1,3 @@
+def book_titles(names):
+    for name in names:
+        print(name.title())