+++ codewof/programming/content/en/print-squares/solution.py
@@ -1,0 +1,3 @@
+def print_squares(numbers):
+    for num in numbers:
+        print(num * num)