+++ codewof/tests/users/test_urls.py
@@ -1,11 +1,10 @@
 import pytest
-from django.conf import settings
 from django.urls import reverse, resolve
 
 pytestmark = pytest.mark.django_db
 
 
-def test_detail(user: settings.AUTH_USER_MODEL):
+def test_detail():
     assert (
         reverse("users:dashboard")
         == f"/users/dashboard/"