+++ codewof/users/admin.py
@@ -3,12 +3,12 @@
 from django.contrib import admin
 from django.contrib.auth import admin as auth_admin
 from django.contrib.auth import get_user_model
+from users.models import UserType
 from users.forms import UserAdminChangeForm, UserAdminCreationForm
 
 User = get_user_model()
 
 
-@admin.register(User)
 class UserAdmin(auth_admin.UserAdmin):
     """Custom user admin class."""
 
@@ -21,3 +21,12 @@ class UserAdmin(auth_admin.UserAdmin):
         'user_type',
         'is_superuser',
     ]
+    ordering = [
+        'first_name',
+        'last_name',
+        'email',
+    ]
+
+
+admin.site.register(User, UserAdmin)
+admin.site.register(UserType)