+++ codewof/users/management/commands/load_user_types.py
@@ -1,8 +1,9 @@
 """Module for the custom Django load_user_types command."""
 
+from copy import deepcopy
 from django.core.management.base import BaseCommand
-from users.models import UserType
 from django.utils.translation import gettext as _
+from users.models import UserType
 
 USER_TYPES = [
     {
@@ -27,7 +28,7 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         """Automatically called when the load_user_types command is given."""
-        for order_number, user_type_data in enumerate(USER_TYPES):
+        for order_number, user_type_data in enumerate(deepcopy(USER_TYPES)):
             slug = user_type_data.pop('slug')
             user_type_data['order'] = order_number
             user_type, created = UserType.objects.update_or_create(